Continue reorganize python scripts into modules.
That was no fun resolving circular references between two import statements.
